# core-lending-credit-cards-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-credit-cards
  description: Credit Cards Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/credit-cards
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}/transactions/{ccTransactionId}:
    get:
      tags:
        - CcTransaction
      summary: Get a transaction record by ID
      operationId: getById_1
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccTransactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcTransactionDTO'
    put:
      tags:
        - CcTransaction
      summary: Update a transaction record
      operationId: update_1
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccTransactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcTransactionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcTransactionDTO'
    delete:
      tags:
        - CcTransaction
      summary: Delete a transaction record
      operationId: delete_1
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccTransactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}/transactions:
    get:
      tags:
        - CcTransaction
      summary: List or search transactions for a revolve line
      operationId: findAll_1
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.ccRevolvingLineId
          in: query
          description: Exact filter for cc revolving line id
          required: false
          schema:
            type: string
        - name: filters.transactionId
          in: query
          description: Exact filter for transaction id
          required: false
          schema:
            type: string
        - name: filters.transactionAmount
          in: query
          description: Exact filter for transaction amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[transactionAmount].from
          in: query
          description: Filter transaction amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[transactionAmount].to
          in: query
          description: Filter transaction amount to value
          required: false
          schema:
            type: string
        - name: filters.currencyCode
          in: query
          description: Exact filter for currency code
          required: false
          schema:
            type: string
            enum:
              - EUR
              - USD
              - GBP
              - CHF
        - name: filters.transactionDate
          in: query
          description: Exact filter for transaction date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[transactionDate].from
          in: query
          description: Filter transaction date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[transactionDate].to
          in: query
          description: Filter transaction date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.transactionType
          in: query
          description: Exact filter for transaction type
          required: false
          schema:
            type: string
            enum:
              - PURCHASE
              - CASH_ADVANCE
              - FEE
              - INTEREST
              - REFUND
              - CREDIT
              - PAYMENT
        - name: filters.transactionStatus
          in: query
          description: Exact filter for transaction status
          required: false
          schema:
            type: string
            enum:
              - AUTHORIZED
              - POSTED
              - REVERSED
              - DECLINED
              - PENDING
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCcTransactionDTO'
    post:
      tags:
        - CcTransaction
      summary: Create a new transaction record
      operationId: create_1
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcTransactionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcTransactionDTO'
...
components:
  schemas:
...
    CcTransactionDTO:
      required:
        - ccRevolvingLineId
        - currencyCode
        - transactionAmount
        - transactionDate
        - transactionStatus
        - transactionType
      type: object
      properties:
        ccTransactionId:
          type: string
          format: uuid
          readOnly: true
        ccRevolvingLineId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        transactionAmount:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: -999999999.99
          exclusiveMinimum: false
          type: number
        currencyCode:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
        transactionDate:
          type: string
          format: date-time
        transactionType:
          type: string
          enum:
            - PURCHASE
            - CASH_ADVANCE
            - FEE
            - INTEREST
            - REFUND
            - CREDIT
            - PAYMENT
        transactionStatus:
          type: string
          enum:
            - AUTHORIZED
            - POSTED
            - REVERSED
            - DECLINED
            - PENDING
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseCcTransactionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CcTransactionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...